Add files via upload
This is me playing to a pattern on my TD-20, for which I added a drums, a guitar, and a keyboard track, just because I love the pattern. It's quite over done on the keyboards and the guitar is as good of sounding "virtual guitar" as I've been able to find or edit.